Address Race Condition When Using PayPal's SDK
It's possible that our modules that use the PayPal SDK will be
initialized prior to `window.paypal` being available on the page, now
that the JS is loaded in asynchronously, causing unnecessary loss in
functionality. Additionally, there doesn't seem to be a way to bind to
any kind of event given off by the SDK on `window` that would allow us
to know whether the SDK is truly loaded or not. To address this, when
the `window.paypal` object is not available yet, and there's a
`<script>` tag on the page loading the PayPal SDK (indicating that we
are indeed expecting the SDK to be there), the PayPal modules will
re-try initializing every 5 seconds until `window.paypal` is available.

PAYPAL-7